Split up DogsList to presentational (DogsList to show data) and container (DogsListContainer to fetch data from API) components, DogsList can be a simple function component
