+++ src/App.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import './App.css';
-import DogsList from './components/DogsList';
+import DogsListContainer from './components/DogsListContainer';
 
 function App() {
   return (
     <div className="App">
-      <DogsList />
+      <DogsListContainer />
     </div>
   );
 }
